+++ lib/cranky/job.rb
@@ -8,6 +8,7 @@ module Cranky
       @defaults = {}
       @target = target
       @overrides = overrides
+      @return_attributes = overrides.delete(:_return_attributes)
     end
 
     def attributes
@@ -19,25 +20,43 @@ module Cranky
       @defaults = defs
     end
 
-    def execute
-      item = get_constant(attributes[:class] ? attributes[:class] : @target).new
-      # Assign all explicit attributes first
-      attributes.each do |attribute, value|
-        unless value == :skip
-          item.send("#{attribute}=", value) if item.respond_to?("#{attribute}=") && !value.respond_to?("call")
+    # Returns the created item
+    def item
+      return @item if @item
+      if @return_attributes
+        @item = Hash.new
+      else
+        @item = get_constant(attributes[:class] ? attributes.delete(:class) : @target).new
+      end
+    end
+
+    # Assign the value to the given attribute of the item
+    def assign(attribute, value)
+      unless value == :skip || attribute == :class
+        if item.respond_to?("#{attribute}=")
+          item.send("#{attribute}=", value)
+        elsif item.is_a?(Hash)
+          item[attribute] = value
         end
       end
-      # Then call any blocks
-      attributes.each do |attribute, value|
-        item.send("#{attribute}=", value.call(item)) if item.respond_to?("#{attribute}=") && value.respond_to?("call")
-      end
+    end
+
+    def execute
+      values = attributes.reject { |attribute, value| value.respond_to?("call") }
+      blocks = attributes.select { |attribute, value| value.respond_to?("call") }
+
+      values.each { |attribute, value| assign(attribute, value) }
+      blocks.each { |attribute, value| assign(attribute, value.call(*(value.arity > 0 ? [item] : []))) }
+
       item
     end
 
     private
 
-      # Nicked from here: http://gist.github.com/301173 
+      # Nicked from here: http://gist.github.com/301173
       def get_constant(name_sym)
+        return name_sym if name_sym.is_a? Class
+
         name = name_sym.to_s.split('_').collect {|s| s.capitalize }.join('')
         Object.const_defined?(name) ? Object.const_get(name) : Object.const_missing(name)
       end